requirements.txt
astroid==2.4.2
colorama==0.4.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.5
opencv-contrib-python==4.2.0.34
Pillow==7.1.2
pylint==2.5.3
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
wrapt==1.12.1
